% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{code64_automation}
\alias{code64_automation}
\title{Code 64 Email Automation}
\usage{
code64_automation(.delete_file = FALSE, .email)
}
\arguments{
\item{.delete_file}{Default is FALSE, TRUE will delete file.}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
Run the \code{\link[=code64_charged_accounts_query]{code64_charged_accounts_query()}} and email them to Performance Improvement
and the Medical Chairperson

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Will ask you where you want to save the file so you can email it out, will
delete the file upon email send.
\item Uses the \code{\link[=code64_charged_accounts_query]{code64_charged_accounts_query()}} internally.
}
}
\examples{
\dontrun{
library(RDCOMClient)
code64_automation(.email = "someone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
