% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{code64_charged_accounts_query}
\alias{code64_charged_accounts_query}
\title{Code 64 Charged Accounts}
\usage{
code64_charged_accounts_query()
}
\value{
A tibble
}
\description{
Get accounts that have a code 64 charged to them
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has Start and End dates set dynamically in the query in order to obtain
for the previous month. For example if you run the query on any day in October
you will get data for September.
}
}
\examples{
library(DBI)
library(dplyr)
library(janitor)
code64_charged_accounts_query()

}
\author{
Steven P. Sanderson II, MPH
}
