% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_consultant_report_functions.R
\name{coded_consults_query}
\alias{coded_consults_query}
\title{OPPE Coded Consults Query}
\usage{
coded_consults_query(.resp_pty)
}
\arguments{
\item{.resp_pty}{The providers six digit ID Number. This should be quoted like '123456'}
}
\value{
A tibble
}
\description{
Get the coded consults for a specified provider using their ID number or resp_pty cd.
}
\details{
\itemize{
\item Get the coded consults for a specified consulting provider using their ID number/resp_pty cd.
\item This gets data from DSS, you must have a connection file on your pc to use this function.
If you do not have one then submit a help desk ticket for one.
\item Gets data from smsdss.c_Coded_Consults_v for the last 6 full months by discharge date.
\item Uses timetk::tk_augment_timeseries_signature() to obtain a few different time components
}
}
\examples{
library(janitor)
library(dplyr)
library(tibble)
library(DBI)
library(odbc)
library(lubridate)
library(timetk)
library(tidyquant)
coded_consults_query(.resp_pty = '123456')

}
\author{
Steven P. Sanderson II, MPH
}
