% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_consultant_report_functions.R
\name{coded_consults_top_plt}
\alias{coded_consults_top_plt}
\title{OPPE Plot of top N Attending}
\usage{
coded_consults_top_plt(.data)
}
\arguments{
\item{.data}{The data provided the coded_consults_top_providers function}
}
\value{
A ggplot plot
}
\description{
Plot out the results from the coded_consults_top_providers function
}
\details{
\itemize{
\item Gives a ggplot2 plot of the top_n attending providers from the coded_consults_top_providers
function
}
}
\examples{
library(janitor)
library(dplyr)
library(tibble)
library(DBI)
library(odbc)
library(lubridate)
library(timetk)
library(tidyquant)
library(ggplot2)
coded_consults_query(.resp_pty = "013128") \%>\%
  coded_consults_top_providers_tbl(
    .top_n          = 10,
    .attending_col  = attending_md,
    .consultant_col = consultant
  ) \%>\%
  coded_consults_top_plt()

}
\author{
Steven P Sanderson II, MPH
}
