% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_consultant_report_functions.R
\name{coded_consults_top_providers_tbl}
\alias{coded_consults_top_providers_tbl}
\title{OPPE Coded Consults Top Attending}
\usage{
coded_consults_top_providers_tbl(
  .data,
  .top_n,
  .attending_col,
  .consultant_col
)
}
\arguments{
\item{.data}{The data that you provided. There must be an Attending column and Consultant column}

\item{.top_n}{How many of the top attending providers do you want returned.}

\item{.attending_col}{The column that holds the name of the attending provider}

\item{.consultant_col}{The column that holds the name of the consulting provider(s)}
}
\description{
Gets the top n attending providers who have consulted the consultant in question.
}
\details{
\itemize{
\item Get the top n attending providers and counts of times they consulted the specialist
provider in question.
\item This gets data from either DSS by using the \code{\link[=coded_consults_query]{coded_consults_query()}} or an
imported file.
}
}
\examples{
library(dplyr)
library(tibble)
coded_consults_top_providers_tbl(
.data = coded_consults_query(.resp_pty = "013128")
, .top_n = 10
, .attending_col = attending_md
, .consultant_col = consultant
)

}
\author{
Steven P. Sanderson II, MPH
}
