% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_consultant_report_functions.R
\name{coded_consults_trend_plt}
\alias{coded_consults_trend_plt}
\title{OPPE Coded Consults Trend Plot}
\usage{
coded_consults_trend_plt(.data, .consultant_col)
}
\arguments{
\item{.data}{The data that you want to trend}

\item{.consultant_col}{The column that holds the consultant name}
}
\value{
A ggplot2 plot
}
\description{
Get a plot of the trend for a spcified consultant
}
\details{
\itemize{
\item Get the trend plot for a given consultant
\item Data can come from the coded_consults_query function or data that is imported
\item Must have a consultant column
\item Groups data by month label, January, February, etc.
\item See \code{\link[=coded_consults_query]{coded_consults_query()}}
}
}
\examples{
library(dplyr)
library(tibble)
library(ggplot2)
coded_consults_query(.resp_pty = "013128") \%>\%
  coded_consults_trend_plt(.consultant_col = consultant)

}
\author{
Steven P. Sanderson II, MPH
}
