% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{congenital_malformation_automation}
\alias{congenital_malformation_automation}
\title{Congenital Malformation Automation for HIM}
\usage{
congenital_malformation_automation(.delete_file = FALSE, .email)
}
\arguments{
\item{.delete_file}{Default is FALSE, TRUE will delete file}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This automation will send information on congenital malformations that comes
from the \code{\link[=congenital_malformation_query]{congenital_malformation_query()}} to HIM

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
\itemize{
\item Uses the \code{\link[=congenital_malformation_query]{congenital_malformation_query()}} internally.
\item Requires a connection to DSS and uses \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
}
}
\examples{
\dontrun{
library(RDCOMClient)
congenital_malformation_automation(.email= "someone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
