% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{congenital_malformation_query}
\alias{congenital_malformation_query}
\title{Congenital Malformations Query for HIM}
\usage{
congenital_malformation_query()
}
\description{
This query will get the information for congenital malformations that must
be sent to HIM on a monthly basis.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically
}
}
\examples{
\dontrun{
congenital_malformation_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
