% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_obj.R
\name{db_connect}
\alias{db_connect}
\title{DSS DB Connection Function}
\usage{
db_connect()
}
\value{
A SQL-SERVER DSS connection object
}
\description{
Connect to DSS
}
\details{
\itemize{
\item DSS Connection function
\item Must have a valid DSS connection object on your local pc and DSS permissions
}
}
\examples{
db_connect()

}
\author{
Steven P. Sanderson II, MPH
}
