% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_obj.R
\name{db_disconnect}
\alias{db_disconnect}
\title{DSS DB Disconnect Function}
\usage{
db_disconnect(.connection)
}
\arguments{
\item{.connection}{The connection object returned from \code{\link[=db_connect]{db_connect()}}}
}
\value{
A SQL-SERVER DSS connection object
}
\description{
Disconnect from DSS
}
\details{
\itemize{
\item DSS Disconnect function
\item Must have a valid DSS connection object on your local pc and DSS permissions
}
}
\examples{
db_disconnect()

}
\author{
Steven P. Sanderson II, MPH
}
