% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{denials_admits_by_ed_query}
\alias{denials_admits_by_ed_query}
\title{Denials Admits by ED}
\usage{
denials_admits_by_ed_query(.name)
}
\arguments{
\item{.name}{The name of the provider you want returned}
}
\value{
A tibble
}
\description{
This query gets the admit counts by ed provider
}
\details{
\itemize{
\item Queries DSS the linked WellSoft Report DB
\item Uses the \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}} functions
\item Does not need an argument but one can be passes til filter on name if desired,
although this should be done as a dplyr::filter() so this funcationality
may be taken away in the future
}
}
\examples{
denials_admits_by_ed_query()

}
\author{
Steven P. Sanderson II, MPH
}
