% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{denials_admits_by_md_query}
\alias{denials_admits_by_md_query}
\title{Denials Admits by MD}
\usage{
denials_admits_by_md_query(.name)
}
\arguments{
\item{.name}{The name of the provider you want returned, leave empty for all}
}
\value{
A Tibble
}
\description{
This query gets the admit counts by attending provider
}
\details{
\itemize{
\item Queries DSS
\item Uses the \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}} functions
\item Does not need an argument but one can be passed to filter on name if
desired, although this should be done as a dplyr::filter() so this funcationality
may be taken away in the future
\item The data is grouped by:
}
\enumerate{
\item Payer_Category
\item Attending Dr
\item Admit Month
\item Admit Year
\item Primary Payer Plan Code
\item Med Staff Department
}
}
\examples{
library(DBI)
library(dplyr)
library(tibble)
library(janitor)
library(data.table)

denials_admits_by_md_query()

}
\author{
Steven P. Sanderson II, MPH
}
