% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{discharge_order_to_discharge_automation}
\alias{discharge_order_to_discharge_automation}
\title{Discharge Order to Discharge Automation}
\usage{
discharge_order_to_discharge_automation(.delete_file = FALSE, .email)
}
\arguments{
\item{.delete_file}{Default is FALSE, TRUE will delete file}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
Uses the \code{\link[=discharge_order_to_discharge_query]{discharge_order_to_discharge_query()}} to get information from DSS
for processing.

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
\itemize{
\item Gets the last discharge order written for a visit and compares that with the
time that was input as the discharge date time
\item Uses the \code{\link[=discharge_order_to_discharge_query]{discharge_order_to_discharge_query()}} internally to get the data.
}
}
\examples{
\dontrun{
library(RDCOMClient)
discharge_order_to_discharge_automation(.email = "someone@gmail.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
