% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{discharge_order_to_discharge_query}
\alias{discharge_order_to_discharge_query}
\title{Discharge Order to Discharge Date Time Query}
\usage{
discharge_order_to_discharge_query()
}
\value{
A tibble
}
\description{
This query will get information on the last discharge order written for a patient
and the time that was input into the system Soarian/Invision as the discharge date time.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically
}
}
\examples{
\dontrun{
discharge_order_to_discharge_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
