% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{duplicate_coded_cataract_query}
\alias{duplicate_coded_cataract_query}
\title{Duplicate Coded Cataracts Query for HIM}
\usage{
duplicate_coded_cataract_query()
}
\value{
A tibble
}
\description{
This query will get the accounts that have been duplicate coded for a cataract
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically
}
}
\examples{
\dontrun{
duplicate_coded_cataract_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
