% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{geocode_discharges_query}
\alias{geocode_discharges_query}
\title{Get Address for discharges that can be geocoded}
\usage{
geocode_discharges_query()
}
\value{
A tibble
}
\description{
Get discharged accounts from DSS
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item This will look back at discharges starting with a discharge date of six months prior
to the SQL GETEDATE() function
}
}
\examples{
library(DBI)
library(tibble)

geocode_discharges_query()

}
\author{
Steven P. Sanderson II, MPH
}
