% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{infection_prevention_patient_days_query}
\alias{infection_prevention_patient_days_query}
\title{Patient Days for Infection Prevention Query}
\usage{
infection_prevention_patient_days_query()
}
\value{
A tibble
}
\description{
This gets the data necessary for infection prevention. This is used for a variety
of reasons, one being total vent days against total days.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has Start and End dates set dynamically in the query in order to obtain data
for the previous month.
}
}
\examples{
\dontrun{
infection_prevention_patient_days_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
