% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data_mainpulation_tbls.R
\name{infection_prevention_patient_days_tbl}
\alias{infection_prevention_patient_days_tbl}
\title{Patient Days for Infection Prevention Tibble}
\usage{
infection_prevention_patient_days_tbl(.data)
}
\arguments{
\item{.data}{The data that results from \code{\link[=infection_prevention_patient_days_query]{infection_prevention_patient_days_query()}}}
}
\value{
An excel workbook object pointer
}
\description{
This takes data in from the \code{\link[=infection_prevention_patient_days_query]{infection_prevention_patient_days_query()}} and performs the
necessary manipulation/cleanup to get it ready to email out or use in conjunction
with \code{\link[=save_to_excel]{save_to_excel()}}
}
\details{
\itemize{
\item Returns a tibble
\item Expects data from the \code{\link[=infection_prevention_patient_days_query]{infection_prevention_patient_days_query()}}
}
}
\examples{
\dontrun{
library(tidyverse)

infection_prevention_patient_days_query() \%>\%
  infection_prevention_patient_days_tbl()
}

}
\author{
Steven P. Sanderson II, MPH
}
