% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{inpatient_coding_lag_automation}
\alias{inpatient_coding_lag_automation}
\title{Inpatient Coding Lag Automation}
\usage{
inpatient_coding_lag_automation(.data, .delete_file = FALSE, .email)
}
\arguments{
\item{.data}{The data that comes in typically from the \code{\link[=inpatient_coding_lag_tbl]{inpatient_coding_lag_tbl()}}}

\item{.delete_file}{Default is FALSE, TRUE will delete file}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This takes in data from the function \code{\link[=inpatient_coding_lag_tbl]{inpatient_coding_lag_tbl()}}

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
\itemize{
\item This function will ask you where you want to save the data so that it can be emailed out
\item After the email is sent the function will delete the saved file
}
}
\examples{
\dontrun{
library(RDCOMClient)
library(tidyverse)
inpatient_coding_lag_query() \%>\%
  inpatient_coding_lag_tbl() \%>\%
  inpatient_coding_lag_automation(.email = "someone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
