% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{inpatient_coding_lag_query}
\alias{inpatient_coding_lag_query}
\title{Inpatient Coding Lag Query for HIM}
\usage{
inpatient_coding_lag_query()
}
\value{
A tibble
}
\description{
This query will get results for the inpatient coding lag report for HIM
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically
}
}
\examples{
\dontrun{
inpatient_coding_lag_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
