% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data_mainpulation_tbls.R
\name{inpatient_coding_lag_tbl}
\alias{inpatient_coding_lag_tbl}
\title{Inpatient Coding Lag tibble}
\usage{
inpatient_coding_lag_tbl(.data)
}
\arguments{
\item{.data}{Data that you want to manipulate typically from \code{\link[=inpatient_coding_lag_query]{inpatient_coding_lag_query()}}}
}
\value{
A tibble
}
\description{
This function takes in the query \code{\link[=inpatient_coding_lag_query]{inpatient_coding_lag_query()}} as the data
argument and augments the data so that it can be taken into the
\code{\link[=inpatient_coding_lag_automation]{inpatient_coding_lag_automation()}} function.
}
\details{
\itemize{
\item Takes in \code{\link[=inpatient_coding_lag_query]{inpatient_coding_lag_query()}} as the data argument.
\item Returns a tibble that can be used in the \code{\link[=inpatient_coding_lag_automation]{inpatient_coding_lag_automation()}}
function.
}
}
\examples{
\dontrun{
library(tidyverse)

inpatient_coding_lag_query() \%>\%
  inpatient_coding_lag_tbl()
}

}
\author{
Steven P. Sanderson II, MPH
}
