% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_tbl_funtions.R
\name{kmeans_mapped_tbl}
\alias{kmeans_mapped_tbl}
\title{OPPE CPOE K-Means Mapper}
\usage{
kmeans_mapped_tbl(.data, .centers = 15)
}
\arguments{
\item{.data}{You must have a tibble in the working environment that is produced
by using the following work flow: oppe_cpoe_query() \%>\% oppe_cpoe_tbl() \%>\%
oppe_cpoe_user_item_tbl()}

\item{.centers}{How many different centers do you want to try}
}
\value{
A nested tibble
}
\description{
Create a tibble that maps the \code{\link[=oppe_cpoe_kmeans]{oppe_cpoe_kmeans()}} using [purrr::map()
to create a nested data.frame/tibble that holds n centers. This tibble will be
used to help create a scree plot.
}
\details{
Takes in a single parameter of .centers. This is used to create the tibble
and map the \code{\link[=oppe_cpoe_kmeans]{oppe_cpoe_kmeans()}} function down the list creating a nested tibble.
}
\examples{
ui_tbl <- oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl() \%>\%
  oppe_cpoe_user_item_tbl()

kmeans_mapped_tbl(ui_tbl)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Scree_plot}
}
\author{
Steven P. Sanderson II, MPH
}
