% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_tbl_funtions.R
\name{kmeans_tidy}
\alias{kmeans_tidy}
\title{K-Means tidy Functions}
\usage{
kmeans_tidy(.kmeans_obj, .tidy_type = "tidy")
}
\arguments{
\item{.kmeans_obj}{A \code{\link[stats:kmeans]{stats::kmeans()}} object}

\item{.tidy_type}{"tidy","glance", or "augment"}
}
\value{
A tibble
}
\description{
K-Means tidy functions
}
\details{
Takes in a k-means object and returns one of the items asked
for. Either the \code{\link[broom:reexports]{broom::tidy()}} \code{\link[broom:reexports]{broom::glance()}} or \code{\link[broom:reexports]{broom::augment()}}. The
function defaults to \code{\link[broom:reexports]{broom::tidy()}}
}
\examples{
oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl() \%>\%
  oppe_cpoe_user_item_tbl() \%>\%
  oppe_cpoe_kmeans() \%>\%
  kmeans_tidy()

}
\author{
Steven P. Sanderson II, MPH
}
