% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los_ra_index.R
\name{los_ra_index_plt}
\alias{los_ra_index_plt}
\title{Plot LOS and Readmit Index with Variance}
\usage{
los_ra_index_plt(.data)
}
\arguments{
\item{.data}{The data supplied from \code{\link[=los_ra_index_summary_tbl]{los_ra_index_summary_tbl()}}}
}
\value{
A patchwork ggplot2 plot
}
\description{
Plot the index of the length of stay and readmit rate against each other along
with the variance
}
\details{
\itemize{
\item Expects a tibble
\item Expects a Length of Stay and Readmit column, must be numeric
\item Uses patchwork to stack plots
}
}
\examples{
library(dplyr)
library(ggplot2)
library(patchwork)
library(tidyquant)

los_ra_index_query() \%>\%
los_ra_index_summary_tbl() \%>\%
los_ra_index_plt()

}
\author{
Steven P. Sanderson II, MPH
}
