% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los_ra_index.R
\name{los_ra_index_summary_tbl}
\alias{los_ra_index_summary_tbl}
\title{Make LOS and Readmit Index Summary Tibble}
\usage{
los_ra_index_summary_tbl(.data)
}
\arguments{
\item{.data}{The data you are going to analyze from \code{\link[=los_ra_index_query]{los_ra_index_query()}}}
}
\value{
A tibble
}
\description{
Create the length of stay and readmit index summary tibble
}
\details{
\itemize{
\item Expects a tibble
\item Adds a year column from the Admit Date and one from Discharge Date
\item Uses all data to compute variance, if you want it for a particular time frame
you will have to filter the data from \code{\link[=los_ra_index_query]{los_ra_index_query()}}
}
}
\examples{
los_ra_index_summary_tbl(
  .data = los_ra_index_query()
)

}
\author{
Steven P. Sanderson II, MPH
}
