% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{monthly_discharge_trauma_query}
\alias{monthly_discharge_trauma_query}
\title{Monthly Discharge Trauma File Query}
\usage{
monthly_discharge_trauma_query()
}
\value{
A tibble
}
\description{
This query is used to grab all possible Trauma related accounts. There are two
queries that are run, one for accounts by admit_date and one for accounts
by discharge_date.

There are no parameters to this function as of yet, it may be introduced in the
future if there is a strong need for it. The dates are set dynamically in the
SQL so that data is always run for the previous third month in which the query is run.
For example if the query is run in any day of November 2020 then the data will
be pulled for August of 2020
}
\details{
\itemize{
\item Need a valid DSS connection and rights to query.
\item Utilizes both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}} functions
\item Returns data as a tibble
}
}
\examples{
\dontrun{
monthly_discharge_trauma_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
