% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{monthly_psy_admits_discharges_automation}
\alias{monthly_psy_admits_discharges_automation}
\title{Monthly PSY Admit and Discharge Automation}
\usage{
monthly_psy_admits_discharges_automation(.data, .delete_file = FALSE, .email)
}
\arguments{
\item{.data}{The output from the tbl function \code{\link[=monthly_psy_admits_discharges_tbl]{monthly_psy_admits_discharges_tbl()}}}

\item{.delete_file}{Default is FALSE, TRUE will delete file}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This takes in the data from \code{\link[=monthly_psy_admits_query]{monthly_psy_admits_query()}} and \code{\link[=monthly_psy_discharges_query]{monthly_psy_discharges_query()}}
puts them in a single excel workbook on different sheets and sends.

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
Takes the output from \code{\link[=monthly_psy_admits_discharges_tbl]{monthly_psy_admits_discharges_tbl()}} and saves it to
an excel workbook using the \code{\link[xlsx:Workbook]{xlsx::saveWorkbook()}} function. The file is saved
to a location specified by the user and upon successful function completion the
file is deleted.
}
\examples{
\dontrun{
library(RDCOMClient)
library(tidyverse)
monthly_psy_admits_discharges_tbl() \%>\%
  monthly_psy_admits_discharges_automation(.email = "someone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
