% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data_mainpulation_tbls.R
\name{monthly_psy_admits_discharges_tbl}
\alias{monthly_psy_admits_discharges_tbl}
\title{Monthly PSY Admit/Discharge tibble}
\usage{
monthly_psy_admits_discharges_tbl()
}
\value{
A java pointer to an excel workbook (or a list in the future)
}
\description{
Takes data from \code{\link[=monthly_psy_admits_query]{monthly_psy_admits_query()}} and \code{\link[=monthly_psy_discharges_query]{monthly_psy_discharges_query()}}
and puts them inside of an excel workbook with data on different sheets.

There are no parameters to this function, it takes in it directly the queries
internally.

The \code{\link[=monthly_psy_admits_discharges_automation]{monthly_psy_admits_discharges_automation()}} will internally handle this.
}
\details{
\itemize{
\item Creates a sheet for the raw admit and discharge data
\item Creates a sheet for distinct Admitted and Discharged MRN's
\item Creates a sheet with pivot tables for Admited/Discharged financial grouping
(i.e. Medicare, Blue Cross)
\item Creates a sheet a count of distinct Admitted and Discharged MRN's
}
}
\examples{
\dontrun{
monthly_psy_admits_discharges_tbl()
}

}
\author{
Steven P. Sanderson II, MPH
}
