% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{monthly_psy_discharges_query}
\alias{monthly_psy_discharges_query}
\title{Monthly PSY Discharges Query}
\usage{
monthly_psy_discharges_query()
}
\value{
A tibble
}
\description{
This query is used to grab all hospital service PSY discharges for the previous
month. There are no parameters to this function as of yet, it may be introduce
in the future if there is a strong need for it.

This information is used to send to Elizabeth Saporito
}
\details{
\itemize{
\item Need a valid DSS connection and rights to query.
\item Utilizes both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}} functions
\item Returns data as a tibble
}
}
\examples{
\dontrun{
monthly_psy_discharges_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
