% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{monthly_trauma_automation}
\alias{monthly_trauma_automation}
\title{Monthly Trauma Automation}
\usage{
monthly_trauma_automation(.data, .delete_file = FALSE, .email)
}
\arguments{
\item{.data}{The output from the tbl function \code{\link[=monthly_trauma_tbl]{monthly_trauma_tbl()}}}

\item{.delete_file}{Default is FALSE, TRUE will delete file}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This takes in the data from \code{\link[=monthly_trauma_tbl]{monthly_trauma_tbl()}} and puts the results into
and excel file with the admits on one sheet and the discharges on another. The
file is saved to a location specified by the user and upon successful completion
the file is deleted.

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\examples{
\dontrun{
library(RDCOMClient)
library(tidyverse)
monthly_trauma_tbl() \%>\%
  monthly_trauma_automation(.email = "someone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
