% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data_mainpulation_tbls.R
\name{monthly_trauma_tbl}
\alias{monthly_trauma_tbl}
\title{Monthly Trauma Tibble}
\usage{
monthly_trauma_tbl()
}
\value{
A java pointer to an excel workbook (or a list in the future)
}
\description{
Takes data from the queries
\enumerate{
\item \code{\link[=monthly_admit_trauma_query]{monthly_admit_trauma_query()}}, and
\item \code{\link[=monthly_discharge_trauma_query]{monthly_discharge_trauma_query()}}
And then performs the necessary transformations on them to get them into one
excel file with two sheets. One for Admits and one for Discharges.
}
}
\details{
\enumerate{
\item Internally gets data from \code{\link[=monthly_admit_trauma_query]{monthly_admit_trauma_query()}}
\item Internally gets data from \code{\link[=monthly_discharge_trauma_query]{monthly_discharge_trauma_query()}}
\item Transforms them into one excel file with two sheets that can be
written out to disk
}
}
\examples{
\dontrun{
monthly_trauma_tbl()
}

}
\author{
Steven P. Sanderson II, MPH
}
