% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{myhealth_monthly_surgery_automation}
\alias{myhealth_monthly_surgery_automation}
\title{MyHealth Monthly Surgery Automation}
\usage{
myhealth_monthly_surgery_automation(.data, .delete_file = FALSE, .email)
}
\arguments{
\item{.data}{The data from \code{\link[=myhealth_monthly_surgery_tbl]{myhealth_monthly_surgery_tbl()}}}

\item{.delete_file}{FALSE is the default, so the file will be kept, TRUE will
delete the file after function completion}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This function will send out the results of the \code{\link[=myhealth_monthly_surgery_tbl]{myhealth_monthly_surgery_tbl()}}

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
\itemize{
\item Takes in the results of \code{\link[=myhealth_monthly_surgery_tbl]{myhealth_monthly_surgery_tbl()}} function
\item Asks user where to save the temporary file
\item Will delete file after sending if parameter is set to true
}
}
\examples{
\dontrun{
library(RDCOMClient)
library(tidyverse)
myhealth_monthly_surgery_query() \%>\%
  myhealth_monthly_surgery_tbl() \%>\%
  myhealth_monthly_surgery_automation(.email = "someone@email.com")
}

}
\author{
Steven P. Sanderson II
}
