% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{myhealth_monthly_surgery_query}
\alias{myhealth_monthly_surgery_query}
\title{MyHealth Monthly Surgery File}
\usage{
myhealth_monthly_surgery_query()
}
\value{
A tibble
}
\description{
This query will generate the data needed to produce the monthly MyHealth Surgery
file and send to MyHealth.
}
\details{
\itemize{
\item Need a valid DSS connection and rights to query.
\item Utilizes both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}} functions
\item Returns data as a tibble
\item This query contains the following fields:
}
\enumerate{
\item Facilty Account Number
\item Provider_ID
\item Provider Short Name
\item Room ID
\item Start Date
\item Enter Procedure Room Time
\item Leave Procedure Room Time
\item Procedure Description (from ORSOS)
\item Anesthesia Start Date
\item Anesthesia Start Time
\item Anesthesia Stop Date
\item Anesthesia Stop Time
\item Patient Type
\item Admit Recovery Date
\item Admit Recovery Time
\item Leave Recovery Date
\item Leave Recovery Time
}

The date fields are dynamically set to look back one month. There are no
parameters to this function
}
\examples{
\dontrun{
  myhealth_monthly_surgery_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
