% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data_mainpulation_tbls.R
\name{myhealth_monthly_surgery_tbl}
\alias{myhealth_monthly_surgery_tbl}
\title{MyHealth Monthly Surgery Tibble}
\usage{
myhealth_monthly_surgery_tbl(.data)
}
\arguments{
\item{.data}{The result of the \code{\link[=myhealth_monthly_surgery_query]{myhealth_monthly_surgery_query()}}}
}
\value{
A tibble
}
\description{
This takes data in from the \code{\link[=myhealth_monthly_surgery_query]{myhealth_monthly_surgery_query()}} and performs the
necessary manipulation/cleanup to get it ready to email out or use in conjuntion
with \code{\link[=save_to_excel]{save_to_excel()}}
}
\details{
\itemize{
\item Returns a tibble
\item Expects data from the \code{\link[=myhealth_monthly_surgery_query]{myhealth_monthly_surgery_query()}}
}
}
\examples{
\dontrun{
library(tidyverse)

myhealth_monthly_surgery_query() \%>\%
  myhealth_monthly_surgery_tbl()
}

}
\author{
Steven P. Sanderson II, MPH
}
