% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_item_list}
\alias{named_item_list}
\title{Named list from a grouped tibble}
\usage{
named_item_list(.data, .grouping_var)
}
\arguments{
\item{.data}{The grouped data.frame/tibble}

\item{.grouping_var}{The column that contains the grouping variable}
}
\description{
Takes in a data.frame/tibble and creates a named list that can be
used in conjunction with something like \code{\link[=save_to_excel]{save_to_excel()}}
}
\details{
\itemize{
\item Needs a grouped data.frame/tibble
}
}
\examples{
\dontrun{
named_item_list(.data = df, .grouping_var = service_line)
}

}
\author{
Steven P. Sanderson II, MPH
}
