% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_plt_funtions.R
\name{oppe_alos_plt}
\alias{oppe_alos_plt}
\title{OPPE ALOS Plots}
\usage{
oppe_alos_plt(.data, .date_col, .by_time = "month")
}
\arguments{
\item{.data}{The data that you would pass from \code{\link[=oppe_alos_tbl]{oppe_alos_tbl()}}}

\item{.date_col}{The column holding the date value}

\item{.by_time}{How you want the data time aggregated, the default is \strong{"month"}}
}
\value{
A patchwork time series plot
}
\description{
Get plots for a single providers length of stay data.
}
\details{
Takes data in from the \code{\link[=oppe_alos_tbl]{oppe_alos_tbl()}} function.
}
\examples{
oppe_alos_query() \%>\%
  oppe_alos_tbl(.provider_id = "017236") \%>\%
  oppe_alos_plt(.date_col = dsch_date)

}
\author{
Steven P. Sanderson II, MPH
}
