% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_query_funtions.R
\name{oppe_alos_query}
\alias{oppe_alos_query}
\title{OPPE ALOS Query}
\usage{
oppe_alos_query()
}
\value{
A tibble object
}
\description{
Get the average length of stay data from DSS in order to run the OPPE report.
The data in this report goes back 18 months and the dates are set dynamically in
the sql.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically to look back on data to ensure it has
gone through the coding process.
\item The fields that come back are:
\itemize{
\item pt_id
\item ptno_num
\item dsch_date
\item dsch_yr
\item dsch_month
\item atn_dr_no
\item atn_dr_name
\item drg_no
\item lihn_service_line
\item hosim
\item apr_drg
\item severity_of_illness
\item los
\item elos
\item threshold
\item outlier_flag
\item drg_cost_weight
\item pyr_grouping
\item case_var
\item case_index
\item index_threshold
\item z_score
\item med_staff_dept
\item ward_cd
}
\item The tables that are used are:
\itemize{
\item smsdss.c_LIHN_Svc_Line_tbl
\item smsdss.BMH_PLM_PtAcct_V
\item Customer.Custom_DRG
\item smsdss.c_LIHN_SPARCS_BenchmarkRates
\item smsdss.pract_dim_v
\item smsdss.c_LIHN_APR_DRG_OutlierThresholds
\item smsdss.pyr_dim_v
\item smsmir.vst_rpt
}
}
}
\examples{
library(dplyr)

oppe_alos_query() \%>\%
glimpse()

}
\author{
Steven P. Sanderson II, MPH
}
