% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_tbl_funtions.R
\name{oppe_cpoe_cluster_trends_tbl}
\alias{oppe_cpoe_cluster_trends_tbl}
\title{OPPE CPOE Cluster Trends Table}
\usage{
oppe_cpoe_cluster_trends_tbl(.data, .umap_data, ...)
}
\arguments{
\item{.data}{The data that needs to be passed to the function from the \code{\link[=oppe_cpoe_trend_tbl]{oppe_cpoe_trend_tbl()}}
function}

\item{.umap_data}{The umap_kmeans_cluster_results_tbl from the \code{\link[=oppe_cpoe_umap]{oppe_cpoe_umap()}}
function}

\item{...}{Any extra columns you want passed to the function. By default the following
are selected:
\itemize{
\item .cluster
\item order_type
}

You can choose the from the following as extra:
\itemize{
\item hospitalist_np_pa_flag
\item spclty_desc
}}
}
\value{
A tibble
}
\description{
This function takes the data from the \code{\link[=oppe_cpoe_trend_tbl]{oppe_cpoe_trend_tbl()}} function
and creates from that data along with the umap_kmeans_cluster_results_tbl from the
\code{\link[=oppe_cpoe_umap]{oppe_cpoe_umap()}} function.
}
\details{
Requires data from the \code{\link[=oppe_cpoe_trend_tbl]{oppe_cpoe_trend_tbl()}} and \code{\link[=oppe_cpoe_umap]{oppe_cpoe_umap()}}
functions. This data is later used to show how a particular provider stacks up
against others in the same cluster.
}
\examples{
q <- oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl()

trend_tbl <- q \%>\%
  oppe_cpoe_trend_tbl()

ui_tbl <- q \%>\%
  oppe_cpoe_user_item_tbl()

kmm_tbl <- ui_tbl \%>\%
  kmeans_mapped_tbl()

umap_obj <- oppe_cpoe_umap(
  .data = ui_tbl
  , .kmeans_map_tbl = kmm_tbl
  , .k_cluster = 3
)

oppe_cpoe_cluster_trends_tbl(trend_tbl, umap_obj)
oppe_cpoe_cluster_trends_tbl(trend_tbl, umap_obj, hospitalist_np_pa_flag)

}
\author{
Steven P. Sanderson II, MPH
}
