% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_tbl_funtions.R
\name{oppe_cpoe_kmeans}
\alias{oppe_cpoe_kmeans}
\title{OPPE CPOE K-Means}
\usage{
oppe_cpoe_kmeans(.data, .centers = 5)
}
\arguments{
\item{.data}{The data that gets passed from \code{\link[=oppe_cpoe_user_item_tbl]{oppe_cpoe_user_item_tbl()}}}

\item{.centers}{How many initial centers to start with}
}
\value{
A stats k-means object
}
\description{
Takes the output of the \code{\link[=oppe_cpoe_user_item_tbl]{oppe_cpoe_user_item_tbl()}} function and applies the
k-means algorithm to it using \code{\link[stats:kmeans]{stats::kmeans()}}
}
\details{
Uses the \code{\link[stats:kmeans]{stats::kmeans()}} function and creates a wrapper around it.
}
\examples{
oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl() \%>\%
  oppe_cpoe_user_item_tbl() \%>\%
  oppe_cpoe_kmeans()

}
\author{
Steven P. Sanderson II, MPH
}
