% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_query_funtions.R
\name{oppe_cpoe_query}
\alias{oppe_cpoe_query}
\title{OPPE CPOE Query}
\usage{
oppe_cpoe_query(.start_date = "")
}
\arguments{
\item{.start_date}{The start date of the data, if null then it will bring back
all data from the table. Uses the \code{\link[timetk:filter_by_time]{timetk::filter_by_time()}} function to filter
the time.}
}
\value{
A tibble object
}
\description{
Get the CPOE data from DSS in order to run the OPPE report.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically to look back on data to ensure it has
gone through the coding process.
\item The fields that come back are:
\itemize{
\item ent_ate
\item req_pty_cd
\item pract_rpt_name
\item spclty_desc
\item hospitalist_np_pa_flag
\item ord_type_abbr
\item specimen_collect
\item written
\item verbal_order
\item communication
\item specimen_redraw
\item cpoe
\item nursing_order
}
\item The tables that are used are:
\itemize{
\item smsdss.c_CPOE_Rpt_Tbl_Rollup_v
\item smsdss.pract_dim_v
}
}
}
\examples{
library(dplyr)

oppe_cpoe_query() \%>\%
glimpse()

}
\author{
Steven P. Sanderson II, MPH
}
