% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_tbl_funtions.R
\name{oppe_cpoe_tbl}
\alias{oppe_cpoe_tbl}
\title{OPPE CPOE Detail Tibble}
\usage{
oppe_cpoe_tbl(.data, .provider_id, .provider_name)
}
\arguments{
\item{.data}{The data to be passed into the tibble function, should be the results
from \code{\link[=oppe_cpoe_query]{oppe_cpoe_query()}}}

\item{.provider_id}{The provider id number you are looking for}

\item{.provider_name}{If you do not know the providers id, then you can use their name
to get a match. You can also use the \code{\link[=pract_dim_v_query]{pract_dim_v_query()}} to find the id number
before using this function, or you can use it inside of the id number parameter. Be
careful with this approach as providers can have the same last name. It is better to
use the function on it's own and select the appropriate id number.
Uses \code{\link[stringr:case]{stringr::str_to_lower()}} so that you can use all lower case letters to
describe the name you are looking for.}
}
\value{
A tibble
}
\description{
Get the CPOE Detail data for use in CPOE for a specified provider.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
}
}
\examples{
\dontrun{
oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl()

oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl(.provider_id = "009142")

oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl(.provider_name = "rakesh")
}

}
\author{
Steven P. Sanderson II, MPH
}
