% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_tbl_funtions.R
\name{oppe_cpoe_trend_tbl}
\alias{oppe_cpoe_trend_tbl}
\title{OPPE CPOE Provider Order Trend Table}
\usage{
oppe_cpoe_trend_tbl(.data)
}
\arguments{
\item{.data}{The data from the workflow of \code{oppe_cpoe_query() \%>\% oppe_cpoe_tbl()}}
}
\value{
A tibble
}
\description{
Gets the oppe cpoe order trend table for all providers
}
\details{
Takes in the data from the \code{\link[=oppe_cpoe_query]{oppe_cpoe_query()}} and \code{\link[=oppe_cpoe_tbl]{oppe_cpoe_tbl()}}
functions.
}
\examples{
oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl() \%>\%
  oppe_cpoe_trend_tbl()

}
\author{
Steven P. Sanderson II, MPH
}
