% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_tbl_funtions.R
\name{oppe_cpoe_umap}
\alias{oppe_cpoe_umap}
\title{OPPE CPOE UMAP Projection}
\usage{
oppe_cpoe_umap(.data, .kmeans_map_tbl, .k_cluster = 5)
}
\arguments{
\item{.data}{The data from the \code{\link[=oppe_cpoe_user_item_tbl]{oppe_cpoe_user_item_tbl()}} function}

\item{.kmeans_map_tbl}{The data from the \code{\link[=kmeans_mapped_tbl]{kmeans_mapped_tbl()}}}

\item{.k_cluster}{Pick the desired amount of clusters from your analysis of the scree plot}
}
\value{
A list of tibbles and the umap object
}
\description{
Create an umap object from the \code{\link[uwot:umap]{uwot::umap()}} function.
}
\details{
This takes in the user item table/matix that is produced by \code{\link[=oppe_cpoe_user_item_tbl]{oppe_cpoe_user_item_tbl()}}
function. This function uses the defaults of \code{\link[uwot:umap]{uwot::umap()}}
}
\examples{
ui_tbl <- oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl() \%>\%
  oppe_cpoe_user_item_tbl()

kmm_tbl <- kmeans_mapped_tbl(ui_tbl)

oppe_cpoe_umap(.data = ui_tbl, kmm_tbl, 3)

}
\seealso{
\itemize{
\item \url{https://cran.r-project.org/package=uwot} (CRAN)
\item \url{https://github.com/jlmelville/uwot} (GitHub)
\item \url{https://github.com/jlmelville/uwot} (arXiv paper)
}
}
\author{
Steven P. Sanderson II, MPH
}
