% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_plt_funtions.R
\name{oppe_cpoe_umap_plt}
\alias{oppe_cpoe_umap_plt}
\title{OPPE CPOE K-Means UMAP Projection Plot}
\usage{
oppe_cpoe_umap_plt(.data, .point_size = 2)
}
\arguments{
\item{.data}{Takes data from the \code{\link[=oppe_cpoe_umap]{oppe_cpoe_umap()}} function. You simply pass
in the list object and it will automatically take the cluster results tibble}

\item{.point_size}{The size of the \code{ggplot2::geom_point()} size.}
}
\description{
Using the \code{uwot} library we visualize the UMAP: Uniform Manifold
Approximation and Projection for Dimension Reduction. \url{https://arxiv.org/abs/1802.03426}

UMAP (Uniform Manifold Approximation and Projection) is a novel manifold
learning technique for dimension reduction. UMAP is constructed from a
theoretical framework based in Riemannian geometry and algebraic topology.
The result is a practical scalable algorithm that applies to real world data.
The UMAP algorithm is competitive with t-SNE for visualization quality, and
arguably preserves more of the global structure with superior run time
performance. Furthermore, UMAP has no computational restrictions on embedding
dimension, making it viable as a general purpose dimension reduction technique for machine learning.
}
\details{
Requires the user item table/matrix that is output from the \code{\link[=oppe_cpoe_user_item_tbl]{oppe_cpoe_user_item_tbl()}}
function.
}
\examples{
ui_tbl <- oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl() \%>\%
  oppe_cpoe_user_item_tbl()

km_map_tbl <- kmeans_mapped_tbl(ui_tbl)
kmeans_scree_plt(km_map_tbl)

umap_obj <- oppe_cpoe_umap(ui_tbl, km_map_tbl, 3)

oppe_cpoe_umap_plt(umap_obj)

}
\author{
Steven P. Sanderson II, MPH
}
