% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_tbl_funtions.R
\name{oppe_cpoe_user_item_tbl}
\alias{oppe_cpoe_user_item_tbl}
\title{OPPE CPOE K-Means}
\usage{
oppe_cpoe_user_item_tbl(.data)
}
\arguments{
\item{.data}{The data that is passed from \code{\link[=oppe_cpoe_tbl]{oppe_cpoe_tbl()}}}
}
\value{
A aggregated/normalized user item tibble
}
\description{
Takes data from the \code{\link[=oppe_cpoe_tbl]{oppe_cpoe_tbl()}} and transforms it into an aggregated/normalized
user-item tibble of order proportions.
}
\details{
This function requires that the \code{\link[=oppe_cpoe_query]{oppe_cpoe_query()}} and \code{\link[=oppe_cpoe_tbl]{oppe_cpoe_tbl()}}
functions be run and takes the output of that work flow as its input. This function
should be used before using a k-mean model. This is commonly referred to as a user_item
matrix because "users" (e.g. providers) tend to be on the rows and "tems" (e.g. orders)
on the columns.

At present this function only selects the following types of orders:
\itemize{
\item Written
\item Verbal
\item Telephone, and
\item CPOE
}
}
\examples{
oppe_cpoe_query() \%>\%
  oppe_cpoe_tbl() \%>\%
  oppe_cpoe_user_item_tbl()

}
\author{
Steven P. Sanderson II, MPH
}
