% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_query_funtions.R
\name{oppe_denials_detail_query}
\alias{oppe_denials_detail_query}
\title{OPPE Denials Detail Query}
\usage{
oppe_denials_detail_query(.provider_id)
}
\arguments{
\item{.provider_id}{The id of the provider you want denial detial for}
}
\value{
A tibble object
}
\description{
Get the Denials Detail data for a specified provider.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically to look back on data to ensure it has
gone through the coding process.
\item The fields that come back are:
*
\item The tables that are used are:
\itemize{
\item smsdss.bmh_plm_ptacct_v
\item smsdss.pract_dim_v
\item smsdss.drg_dim_v
\item smsdss.c_lihn_svc_line_tbl
\item smsdss.c_lihn_op_svc_line_tbl
\item BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.visit_view
\item BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.CTC_VISIT
\item BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.CTC_INSURANCE
\item BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.CTC_UM_Denial
\item BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.CTC_UM_APPEAL
}
}
}
\examples{
\dontrun{
library(dplyr)

oppe_denials_detail_query(.provider_id = "005892") \%>\%
glimpse()

# If provider id is not known you can do either of the following
# pract_dim_v_query(.name = "the providers name") and the pract_no will be
# returned to you, or you could do the following:

oppe_denials_detail_query(
  .provider_id = pract_dim_v_query(.name = "name_here") \%>\%
    dplyr::pull(pract_no)
)
}

}
\author{
Steven P. Sanderson II, MPH
}
