% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oppe_report_query_funtions.R
\name{oppe_readmit_query}
\alias{oppe_readmit_query}
\title{OPPE Readmit Query}
\usage{
oppe_readmit_query()
}
\value{
A tibble object
}
\description{
Get the readmit rate data from DSS in order to run the OPPE report.
The data in this report goes back 18 months and the dates are set dynamically in
the sql.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically to look back on data to ensure it has
gone through the coding process.
\item The fields that come back are:
\itemize{
\item med_rec_no
\item ptno_num
\item adm_date
\item dsch_date
\item payor_category
\item atn_dr_no
\item med_staff_dept
\item lihn_svc_line
\item severity_of_illness
\item dsch_yr
\item dsch_qtr
\item dsch_month
\item dsch_week
\item dsch_day
\item dsch_day_name
\item rpt_month
\item rpt_qtr
\item dsch_disp
\item dsch_disp_desc
\item drg_no
\item drg_cost_weight
\item hospitalist_private
\item hospitalist_private_flag
\item los
\item interim
\item pt_count
\item readmit_count
\item bench_yr
\item readmit_rate_bench
\item ward_cd
\item z_score
\item pract_rpt_name
}
\item The tables that are used are:
\itemize{
\item smsdss.c_readmit_dashboard_detail_tbl
\item smsdss.c_readmit_dashboard_bench_tbl
\item smsmir.vst_rpt
}
}
}
\examples{
library(dplyr)

oppe_readmit_query() \%>\%
glimpse()

}
\author{
Steven P. Sanderson II, MPH
}
