% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{orsos_j_accounts_automation}
\alias{orsos_j_accounts_automation}
\title{ORSOS J Accounts Automation}
\usage{
orsos_j_accounts_automation(.data, .delete_file = FALSE, .email)
}
\arguments{
\item{.data}{The data from \code{\link[=orsos_j_accounts_query]{orsos_j_accounts_query()}}}

\item{.delete_file}{FALSE is the default, so the file will be kept, TRUE will
delete the file after function completion}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This takes in data directly from the \code{\link[=orsos_j_accounts_query]{orsos_j_accounts_query()}} as the
data argument.

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
\itemize{
\item Takes in the results of \code{\link[=orsos_j_accounts_query]{orsos_j_accounts_query()}} function
\item Asks user where to save the temporary file
\item Will delete file after sending if parameter is set to true
}
}
\examples{
\dontrun{
library(RDCOMClient)
library(tidyverse)
orsos_j_accounts_query() \%>\%
  orsos_j_accounts_automation(.email = "somone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
