% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{orsos_j_accounts_query}
\alias{orsos_j_accounts_query}
\title{ORSOS J Accounts Query}
\usage{
orsos_j_accounts_query()
}
\value{
A tibble
}
\description{
This will pull in all of the "J" accounts from ORSOS in order for the amb surg
department to rectify
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has Start and End dates set dynamically in the query in order to obtain data
for the previous week.
}
}
\examples{
\dontrun{
orsos_j_accounts_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
