% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{orsos_to_sproc_automation}
\alias{orsos_to_sproc_automation}
\title{ORSOS to SPROC Case Reconcilliation Automation}
\usage{
orsos_to_sproc_automation(.data, .delete_file = FALSE, .email)
}
\arguments{
\item{.data}{The data that is passed from \code{\link[=orsos_to_sproc_tbl]{orsos_to_sproc_tbl()}}}

\item{.delete_file}{Default is FALSE, TRUE will delete file}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This automation uses the \code{\link[=orsos_to_sproc_query]{orsos_to_sproc_query()}} to get data that is then
manipulated to find the cases done by providers in the smsdss.c_ORSOS_Post_Case_Rpt_tbl
but not in smsmir.sproc, meaning they have not for some reason been coded by
HIM.

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
\itemize{
\item Takes data from the \code{\link[=orsos_to_sproc_tbl]{orsos_to_sproc_tbl()}} function, and makes an excel file to save off
for sending
}
}
\examples{
\dontrun{
library(RDCOMClient)
library(tidyverse)
orsos_to_sproc_query() \%>\%
  orsos_to_sproc_tbl() \%>\%
  orsos_to_sproc_automation(.email = "someone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
