% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{orsos_to_sproc_query}
\alias{orsos_to_sproc_query}
\title{ORSOS to SPROC Case Reconcilliation Query}
\usage{
orsos_to_sproc_query()
}
\value{
A tibble
}
\description{
This function gets a list of accounts from smsdss.c_ORSOS_Post_Case_Rpt_Tbl
and then looks for the encounter numbers in smsmir.sproc to see if the case
was coded to a provider by HIM.

The start and end dates are set dynamically in order to ensure there is a two
week lag in the looking for cases so that they get through the HIM coding process.

For example if we are running the report on November 5th, 2020 a Thursday, then
the report will pull records from 2020-10-11 00:00:00 through 2020-10-17 23:59:00
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has start and end dates set dynamically to look back on data to ensure it has
gone through the coding process.
}
}
\examples{
orsos_to_sproc_query()

}
\author{
Steven P. Sanderson II, MPH
}
