% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data_mainpulation_tbls.R
\name{orsos_to_sproc_tbl}
\alias{orsos_to_sproc_tbl}
\title{ORSOS to SPROC Tibble}
\usage{
orsos_to_sproc_tbl(.data)
}
\arguments{
\item{.data}{The data from \code{\link[=orsos_to_sproc_query]{orsos_to_sproc_query()}}}
}
\value{
A tibble
}
\description{
This function takes the data from \code{\link[=orsos_to_sproc_query]{orsos_to_sproc_query()}} and performs
the necessary manipulations before using \code{\link[=orsos_to_sproc_automation]{orsos_to_sproc_automation()}}
}
\details{
\itemize{
\item Expects the data from \code{\link[=orsos_to_sproc_query]{orsos_to_sproc_query()}} exactly
}
}
\examples{
\dontrun{
library(tidyverse)

orsos_to_sproc_query() \%>\%
  orsos_to_sproc_tbl()

orsos_to_sproc_query() \%>\%
  orsos_to_sproc_tbl() \%>\%
  save_to_excel(.file_name = "orsos_to_sproc")
}

}
\author{
Steven P. Sanderson II, MPH
}
