% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{pract_dim_v_query}
\alias{pract_dim_v_query}
\title{Get practice dimension view with ID Number and Name}
\usage{
pract_dim_v_query(.name)
}
\arguments{
\item{.name}{Can be null in order to return all providers}
}
\value{
A tibble of provider ID Numbers, Names, Med Staff Department, and Specialty
}
\description{
Get data from the smsdss.pract_dim_v view in DSS
}
\details{
\itemize{
\item Requires a connection to DSS
\item Uses the \code{\link[=db_connect]{db_connect()}} function
\item Data comes back sorted in order of Name
}
}
\examples{
library(DBI)
library(dplyr)
library(tibble)
library(janitor)
library(data.table)

pract_dim_v_query()

pract_dim_v_query(.name = "Abadi")

}
\author{
Steven P. Sanderson II
}
