% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{qec_cdi_automation}
\alias{qec_cdi_automation}
\title{CDI QEC Numbers}
\usage{
qec_cdi_automation(.delete_file = FALSE, .email)
}
\arguments{
\item{.delete_file}{Default is FALSE, TRUE will delete the file.}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
Gather Admit / Discharges counts for CDI QEC dashboard
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Will ask you where you want to save the file so you can email it out
\item Uses the \code{\link[=qec_cdi_query]{qec_cdi_query()}} to automatically set the .data argument internally
\item Email currently goes to PMcKenna at LICommunityHospital.org, should this change
please call Steven Sanderson at 2995 to update the email, and redistribute package.
\item Outlook may ask you to Allow the message to send
}
}
\examples{
\dontrun{
library(RDCOMClient)
library(tidyverse)
qec_cdi_automation(.email = "someone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
