% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{qec_cdi_query}
\alias{qec_cdi_query}
\title{Get CDI QEC Numbers from DSS}
\usage{
qec_cdi_query()
}
\value{
A tibble
}
\description{
Gets the QEC numbers for CDI
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has Start and End dates set dynamically in the query in order to obtain
for the previous month. For example if you run the query on any day in October
you will get data for September.
}
}
\examples{
library(DBI)
library(odbc)
qec_cdi_query()

}
\author{
Steven P. Sanderson II, MPH
}
