% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{readmit_psy_to_psy_query}
\alias{readmit_psy_to_psy_query}
\title{Psych to Psych Readmit Query}
\usage{
readmit_psy_to_psy_query()
}
\value{
A tibble
}
\description{
This query gets the data for psy to psy readmits.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has Start and End dates set dynamically in the query in order to obtain data
for the previous 1 month.
}
}
\examples{
\dontrun{
readmit_psy_to_psy_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
