% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data_mainpulation_tbls.R
\name{readmit_psy_to_psy_tbl}
\alias{readmit_psy_to_psy_tbl}
\title{Readmit Psych To Psych Tibble}
\usage{
readmit_psy_to_psy_tbl(.data)
}
\arguments{
\item{.data}{The data you want to pass, namely \code{\link[=readmit_psy_to_psy_query]{readmit_psy_to_psy_query()}}}
}
\value{
A tibble
}
\description{
This function will take the results of the \code{\link[=readmit_psy_to_psy_query]{readmit_psy_to_psy_query()}} and
ensure that the results are in a \code{tibble}
}
\details{
\itemize{
\item Takes in the data from \code{\link[=readmit_psy_to_psy_query]{readmit_psy_to_psy_query()}}
}
}
\examples{
\dontrun{
readmit_psy_to_psy_query() \%>\%
  readmit_psy_to_psy_tbl()
}

}
\author{
Steven P. Sanderson II, MPH
}
