% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{respiratory_vae_automation}
\alias{respiratory_vae_automation}
\title{Respiratory VAE Automation}
\usage{
respiratory_vae_automation(.data, .delete_file = FALSE, .email)
}
\arguments{
\item{.data}{The output from the tbl function \code{\link[=respiratory_vae_tbl]{respiratory_vae_tbl()}}}

\item{.delete_file}{Default is FALSE, TRUE will delete file}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This will email out the Respiratory VAE file to the email specified
in the function. It is expected that the \code{.data} parameter be filled in with the
results of the \code{\link[=respiratory_vae_tbl]{respiratory_vae_tbl()}} function.

The RDCOMClient Library must be called into the \code{namespace} first with any of the following
\itemize{
\item library(RDCOMClient)
\item require(RDCOMClient)
\item if(!require(pacman)) {install.packages("pacman")}
\item pacman::p_load("RDCOMClient")
}
}
\details{
Takes the output from \code{\link[=respiratory_vae_tbl]{respiratory_vae_tbl()}} and saves it to
an excel workbook using the \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}} function. The file is saved
to a location specified by the user and upon successful function completion the
file is deleted.
}
\examples{
\dontrun{
library(RDCOMClient)
respiratory_vae_query() \%>\%
  respiratory_vae_tbl() \%>\%
  respiratory_vae_automation()
}

}
\author{
Steven P. Sanderson II, MPH
}
