% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{respiratory_vae_query}
\alias{respiratory_vae_query}
\title{Respiratory VAE Query}
\usage{
respiratory_vae_query()
}
\value{
A tibble
}
\description{
This query gets the data necessary to perform the VAE (Ventilator Associated Events)
calculation for the FiO2 and PEEP values for respiratory therapy.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has Start and End dates set dynamically in the query in order to obtain data
for the previous 7 days.
}
}
\examples{
\dontrun{
respiratory_vae_query()
}

}
\seealso{
Infromational Link:
\url{https://www.ahrq.gov/hai/tools/mvp/modules/vae/tool.html}

AHRQ Calculator (Javascript must be enabled):
\url{https://nhsn.cdc.gov/VAECalculator/vaecalc_v7.html}
}
\author{
Steven P. Sanderson II, MPH
}
