% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_string_split.R
\name{sql_left}
\alias{sql_left}
\title{Use SQL LEFT type function in R}
\usage{
sql_left(text, num_char)
}
\arguments{
\item{text}{A piece of text or data you want to take a left part of}

\item{num_char}{How many characters of the text you want to take}
}
\value{
A left sided substring of data
}
\description{
Perform a SQL LEFT() on a piece of data
}
\details{
\itemize{
\item Returns a substring of text from the left side
}
}
\examples{
sql_left("This is some text", 4)

}
\author{
Steven P Sanderson II, MPH
}
