% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_string_split.R
\name{sql_mid}
\alias{sql_mid}
\title{Use SQL SUBSTRING() or Excel MID() type function in R}
\usage{
sql_mid(text, start_num, num_char)
}
\arguments{
\item{text}{A piece of text or data you want to take a substring/mid of}

\item{start_num}{Where you want to start your manipulation}

\item{num_char}{How many characters o fthe text you want to take}
}
\value{
A substring/mid section of text
}
\description{
Perform a SQL SUBSTRING() or Excel MID() type function on a piece of data
}
\details{
\itemize{
\item Returns a substring of text from the substring/mid point chosen
}
}
\examples{
sql_mid("this is some text", 6, 2)

}
\author{
Steven P Sanderson II, MPH
}
