% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_string_split.R
\name{sql_right}
\alias{sql_right}
\title{Use SQL RIGHT type function in R}
\usage{
sql_right(text, num_char)
}
\arguments{
\item{text}{A piece of text or data you want to take a right part of}

\item{num_char}{How many characters of the text you want to take}
}
\value{
A right sided substring of data
}
\description{
Perform a SQL RIGHT() on a piece of data
}
\details{
\itemize{
\item Returns a substring of text from the right side
}
}
\examples{
sql_right("This is some text", 4)

}
\author{
Steven P Sanderson II, MPH
}
