% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_queries.R
\name{ts_alos_elos_query}
\alias{ts_alos_elos_query}
\title{Time Series - ALOS/ELOS Query}
\usage{
ts_alos_elos_query()
}
\value{
A tibble
}
\description{
This query gets the data necessary to do a time series analysis on the excess
length of stay rates.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item The start date is set to 2014-04-01 which is when the Customer.Custom_DRG
table starts
\item The end date is set dynamically to the end of the previous month.
\item It gets data from:
}
\enumerate{
\item smsdss.BMH_PLM_PtAcct_V
\item smsdss.c_LIHN_Svc_Line_tbl
\item Customer.Custom_DRG
\item smsdss.c_LIHN_SPARCS_BenchmarkRates
\item smsdss.pract_dim_v
\item smsdss.c_LIHN_APR_DRG_OutlierThresholds
\item smsdss.pyr_dim_v
}
}
\examples{
\dontrun{
ts_alos_elos_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
