% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_data_manipulation_tbls.R
\name{ts_alos_elos_tbl}
\alias{ts_alos_elos_tbl}
\title{Time Series - ALOS/ELOS Excess Tibble}
\usage{
ts_alos_elos_tbl(.data, .by_time = "month", .date_col)
}
\arguments{
\item{.data}{The data passed in from \code{\link[=ts_alos_elos_query]{ts_alos_elos_query()}}}

\item{.by_time}{The choices are "year", "month", "week", defaults to "month"}

\item{.date_col}{The column containing the date variable of interest}
}
\value{
A tibble with the following columns:
\enumerate{
\item date_col
\item visit_count
\item sum_days
\item sum_exp_days
\item alos
\item elos
\item excess_days
\item avg_excess
}
}
\description{
This takes data from the \code{\link[=ts_alos_elos_query]{ts_alos_elos_query()}} and makes an internal summary table using \code{timetk::summarise_by_time}
with the following possible choices: "year", "month", "week", these are checked
inside of the function, if something else is chose an error will be thrown
and the function will exit. It defaults to "month".
}
\details{
\itemize{
\item Returns a tibble
\item Expects \code{\link[=ts_alos_elos_query]{ts_alos_elos_query()}} as the data argument
}
}
\examples{
\dontrun{
library(healthyR)

data <- ts_alos_elos_query()

data \%>\%
  ts_alos_elos_tbl(.by_time = "year")

data \%>\%
  ts_alos_elos_tbl(.by_time = "month") \%>\%
  ts_plt(.date_col = dsch_date, .value_col = avg_excess)

data \%>\%
  ts_alos_elos_tbl()

data \%>\%
  ts_alos_elos_tbl(.by_time = "week") \%>\%
  save_to_excel(.file_name = "weekly_alos_excess")
}

}
\author{
Steven P. Sanderson II, MPH
}
