% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_queries.R
\name{ts_ed_arrivals_query}
\alias{ts_ed_arrivals_query}
\title{Time Series - ER Arrivals Query}
\usage{
ts_ed_arrivals_query()
}
\value{
A tibble
}
\description{
This query gets the data necessary to do a time series analysis on emergency
room arrivals.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item The start date is set to 2010-01-01
\item The end date is set dynamically to the end of the previous month.
\item It gets data from the WellSoft reporting server
\item The data comes back aggregated by arrival which is granular to the minute.
The following columns come back:
}
\itemize{
\item date_col - date is of datetime, so down to the second
\item value - where value simply equals 1
}
}
\examples{
\dontrun{
ts_ed_arrivals_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
