% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_queries.R
\name{ts_ip_census_los_daily_query}
\alias{ts_ip_census_los_daily_query}
\title{Time Series - Inpatient Census/LOS by Day}
\usage{
ts_ip_census_los_daily_query()
}
\value{
A tibble object
}
\description{
Sometimes it is important to know what the census was on any given day, or what
the average length of stay is on any given day, including for those patients
that are not yet discharged. This can be easily achieved.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item The start date is set to 2001-01-01
\item The end date is not set in order to capture those that are still here.
\item It gets data from smsdss.BMH_PLM_PtAcct_V
}
}
\examples{
\dontrun{
ts_ip_census_los_daily_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
