% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_queries.R
\name{ts_op_visits_query}
\alias{ts_op_visits_query}
\title{Time Series - Outpatient Visits Query}
\usage{
ts_op_visits_query()
}
\value{
A tibble
}
\description{
This query gets the data necessary to do a time series analysis on outpatient visits.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item The start date is set to 2001-01-01
\item The end date is set dynamically to the end of the previous month.
\item It gets data from smsdss.BMH_PLM_PtAcct_V
\item The data comes back aggregated by day, using the discharge date which is the same
as the arrival date. The columns that are returned are:
}
\itemize{
\item date_col - date is of date
\item value - where value count of visits that day
}
\itemize{
\item This query excludes unitized and ED visits and those that start with 9
}
}
\examples{
\dontrun{
ts_op_visits_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
