% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_data_manipulation_tbls.R
\name{ts_op_visits_tbl}
\alias{ts_op_visits_tbl}
\title{Time Series - ED Arrivals Tibble}
\usage{
ts_op_visits_tbl(.data, .by_time = "month", .date_col)
}
\arguments{
\item{.data}{The data passed in from \code{\link[=ts_op_visits_query]{ts_op_visits_query()}}}

\item{.by_time}{The choices are "year", "month", "week", "day" defaults to "month"}

\item{.date_col}{The column containing the date variable of interest}
}
\value{
A tibble with the following columns:
\enumerate{
\item date_col
\item value
}
}
\description{
This takes data from the \code{\link[=ts_op_visits_query]{ts_op_visits_query()}} and makes an internal summary table using \code{timetk::summarise_by_time}
with the following possible choices: "year", "month", "week", "day" these are checked
inside of the function, if something else is chose an error will be thrown
and the function will exit. It defaults to "month". This function also uses the
\code{timetk::pad_by_time} function with a value of 0 for time periods that have no data.
}
\details{
\itemize{
\item Returns a tibble
\item Expects \code{\link[=ts_op_visits_query]{ts_op_visits_query()}} as the data argument
}
}
\examples{
\dontrun{
library(healthyR)

data <- ts_op_visits_query()

data \%>\%
  ts_op_visits_tbl(.by_time = "year")

data \%>\%
  ts_op_visits_tbl(.by_time = "month") \%>\%
  ts_plt(.date_col = dsch_date, .value_col = value)

data \%>\%
  ts_op_visits_tbl()

data \%>\%
  ts_op_visits_tbl(.by_time = "week") \%>\%
  save_to_excel(.file_name = "weekly_alos_excess")
}

}
\author{
Steven P. Sanderson II, MPH
}
