% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_control_charts.R
\name{ts_qc_range}
\alias{ts_qc_range}
\title{Time Series Range}
\usage{
ts_qc_range(.value_col)
}
\arguments{
\item{.value_col}{The time series value column, e.g. discharge counts}
}
\value{
A number
}
\description{
Get the range of a given set of numbers.
}
\details{
\itemize{
\item Expects a vector of numbers be supplied
}
}
\examples{
y <- seq(-5:5)
ts_qc_range(y)

}
\author{
Steven P. Sanderson II, MPH
}
