% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_queries.R
\name{ts_readmit_excess_query}
\alias{ts_readmit_excess_query}
\title{Time Series - Readmission Excess Query}
\usage{
ts_readmit_excess_query()
}
\value{
A tibble
}
\description{
This query gets the data necessary to do a time series analysis on the excess
readmission rates.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item The start date is set to 2016-04-01
\item The end date is set dynamically to the end of the previous month.
\item It gets data from:
}
\enumerate{
\item smsdss.BMH_PLM_PtAcct_V
\item smsdss.pract_dim_v
\item Customer.Custom_DRG
\item smsdss.c_LIHN_Svc_Line_Tbl
\item smsdss.vReadmits
\item smsdss.c_Readmit_Dashboard_Bench_Tbl
\item smsdss.c_ppr_apr_drg_global_exclusions
}
}
\examples{
\dontrun{
ts_readmit_excess_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
