% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automations.R
\name{weekly_psy_discharges_automation}
\alias{weekly_psy_discharges_automation}
\title{Weekly PSY Discharges Automation}
\usage{
weekly_psy_discharges_automation(.data, .delete_file = FALSE, .email)
}
\arguments{
\item{.data}{The data from \code{\link[=myhealth_monthly_surgery_tbl]{myhealth_monthly_surgery_tbl()}}}

\item{.delete_file}{FALSE is the default, so the file will be kept, TRUE will
delete the file after function completion}

\item{.email}{Provide the email address for the recipient. The email must be
in double quotes like so: \email{c("person@licommunityhospital.org;person2@licommunityhospital.org")}
using a semi-colon if there is more than one address.}
}
\description{
This takes in data directly from the \code{\link[=weekly_psy_discharges_query]{weekly_psy_discharges_query()}} as the
data argument
}
\details{
\itemize{
\item Will automatically keep the file unless otherwise specified
\item Sends file via email in Outlook
}
}
\examples{
\dontrun{
library(RDCOMClient)
library(tidyverse)
weekly_psy_discharges_query() \%>\%
  weekly_psy_discharges_automation(.email = "someone@email.com")
}

}
\author{
Steven P. Sanderson II, MPH
}
