% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dss_tables.R
\name{weekly_psy_discharges_query}
\alias{weekly_psy_discharges_query}
\title{Weekly PSY Discharges Query}
\usage{
weekly_psy_discharges_query()
}
\value{
A tibble
}
\description{
A query to get the Weekly PSY discharges for the previous week.
}
\details{
\itemize{
\item Requires a connection to DSS, uses both \code{\link[=db_connect]{db_connect()}} and \code{\link[=db_disconnect]{db_disconnect()}}
\item Has Start and End dates set dynamically in the query in order to obtain data
for the previous week.
}
}
\examples{
\dontrun{
weekly_psy_discharges_query()
}

}
\author{
Steven P. Sanderson II, MPH
}
